fix(agent): use effective UID and full uid_t range in preflight
Address Codex review on #166: Linux /proc parsing now reads the
effective UID (field 2) to match attach identity and macOS ps, and
UIDs are Long so values above Int.MAX_VALUE still compare by identity.

+++ agent/src/main/kotlin/dev/sebastiano/spectre/agent/AttachExceptions.kt
@@ -70,8 +70,8 @@ public class AttachPermissionDeniedException(
     targetPid: Long,
     targetUser: String?,
     currentUser: String? = System.getProperty("user.name"),
-    targetUid: Int? = null,
-    currentUid: Int? = null,
+    targetUid: Long? = null,
+    currentUid: Long? = null,
 ) :
     SpectreAttachException(
         buildAttachPermissionDeniedMessage(
@@ -87,8 +87,8 @@ private fun buildAttachPermissionDeniedMessage(
     targetPid: Long,
     targetUser: String?,
     currentUser: String?,
-    targetUid: Int?,
-    currentUid: Int?,
+    targetUid: Long?,
+    currentUid: Long?,
 ): String {
     val targetDesc = formatOwner(targetUser, targetUid)
     val currentDesc = formatOwner(currentUser, currentUid)
@@ -97,7 +97,7 @@ private fun buildAttachPermissionDeniedMessage(
         "the same OS user identity (numeric UID on POSIX when available)."
 }
 
-private fun formatOwner(userName: String?, uid: Int?): String =
+private fun formatOwner(userName: String?, uid: Long?): String =
     when {
         uid != null && userName != null -> "uid=$uid (user '$userName')"
         uid != null -> "uid=$uid"
